+++ programs/mango-v4/src/lib.rs
@@ -30,11 +30,11 @@ pub mod mango_v4 {
 
     use super::*;
 
-    pub fn group_create(ctx: Context<CreateGroup>, group_num: u32, testing: u8) -> Result<()> {
+    pub fn group_create(ctx: Context<GroupCreate>, group_num: u32, testing: u8) -> Result<()> {
         instructions::group_create(ctx, group_num, testing)
     }
 
-    pub fn group_close(ctx: Context<CloseGroup>) -> Result<()> {
+    pub fn group_close(ctx: Context<GroupClose>) -> Result<()> {
         instructions::group_close(ctx)
     }
 
@@ -118,7 +118,7 @@ pub mod mango_v4 {
         instructions::token_deregister(ctx, token_index)
     }
 
-    pub fn token_update_index(ctx: Context<UpdateIndex>) -> Result<()> {
+    pub fn token_update_index(ctx: Context<TokenUpdateIndex>) -> Result<()> {
         instructions::token_update_index(ctx)
     }